perf: Fix training-plans page performance with split hook architecture
Major performance optimization that eliminates multiple API calls and flickering issues on the training-plans page:

- Split useTrainingPlans hook into two focused hooks:
  - useTrainingPlansData: Single-purpose data fetching (used in main page)
  - useTrainingPlansActions: CRUD operations only (used in individual cards)
- Add Axios for better HTTP handling and error management
- Set default theme to dark mode with localStorage persistence
- Remove circular dependencies that caused infinite loops
- Fix TypeScript and linting issues in CoachDashboard component
- Update project documentation with current status and achievements

The training-plans page is now demo-ready with single API call per page load and smooth user experience.
