+++ package.json
@@ -21,6 +21,7 @@
     "@heroui/theme": "^2.4.17",
     "@supabase/supabase-js": "^2.50.5",
     "@tailwindcss/forms": "^0.5.10",
+    "axios": "^1.10.0",
     "bcrypt": "^6.0.0",
     "classnames": "^2.5.1",
     "framer-motion": "^12.23.3",
